Remove confusing key suggestions on invalid keys (#1242)
Currently TF providers may emit error messages that conflate two conditions: a key being passed that is not understood
by the provider (unknown key) and a value being passed to a key that has invalid content. In the second case, the bridge
provides confusing typo-correction suggestions that do not make sense. This is now fixed. Easiest to illustrate by an
example:

Before the change:

  pulumi:providers:aws (default_5_41_0):
    error: pulumi:providers:aws resource 'default_5_41_0' has a problem: could not validate provider configuration:
    Invalid or unknown key. Check `pulumi config get aws:endpoints`. Did you mean `aws:endpoints`?

After the change:

  pulumi:providers:aws (default_5_41_0):
    error: pulumi:providers:aws resource 'default_5_41_0' has a problem: could not validate provider configuration:
    Invalid or unknown key. Check `pulumi config get aws:endpoints`.