Don't panic on computed values in non-string sets in Diff
As part of Diff, we traverse the inputs, which can involve hashing Set values as part of computing nested path shapes for diffs.  However, the Terraform hashing for Set values does not handle unknown/computed values, and instead interprets the unknonw sentinel literally as a string even when being used for a non-string value type.  This leads to a panic.

Instead, we skip hashing set elements that are unknown, and report their diff against the parent.  (Note - I am not 100% sure that is "correct" relative to Terraform behaviour, but cannot find the reference for what Terraform expects to produce in this case).