Add support for computed inputs in diff
This is intended to enable closing:
- pulumi/pulumi-aws#1655

We add a new flag on fields, `ComputedInput`, which indicates that the value
will only appear as a result of a tf diff computation or planning, and not in
`news`. This computed _input_ would ordinarily be computed during `Check()` in
our lifecycle model.

For reasons unclear to me, we don't include computed properties in detailed
diffs, and in stepping through `makeDetailedDiff`, we found that the visitor
never saw properties such as `tags_all.foo` when that property was new and
computed.

With this change, we can mark the `tags_all` field as a `ComputedInput`, and
trigger updates on default tags changing:

```
pulumi up --skip-preview
Updating (test-aws-1655)

View in Browser (Ctrl+O): https://app.pulumi.com/pulumi/simple-yaml/test-aws-1655/updates/9

     Type                     Name                       Status              Info
     pulumi:pulumi:Stack      simple-yaml-test-aws-1655
 ~   ├─ pulumi:providers:aws  aws-provider               updated (0.13s)     [diff: ~defaultTags]
 ~   ├─ aws:s3:Bucket         my-bucket                  updated (0.87s)
 ~   └─ aws:s3:BucketObject   index.html                 updated (0.24s)
```