Change the ElementStrategy interface to accept a *T
This allows applying to an already existing mapping, such as `{Doc: ...}`.

Because it is no longer passable to call a generic strategy and then undo it (since
strategies are now handed a copy of *T), I have changed `Unmappable` to `Ignore`. If a
token is unmapped, it can be caught by our normal mechanism.

This opens up strategies to apply to arbitrary elements of `tfbride.{Resource,DataSource}Info` and to
be layered in sequence.