Finalize expiremental tokens (#1232)
* Copy tokens/x into new package defaults

* Change x.ComputeDefaults to defaults.ApplyStrategy

* Remove explicit Metadata from AutoAliasing

Instead, metadata is picked up from tfbridge.ProviderInfo.

* Add MustApplyStrategy

* Change AutoAliasing to ApplyAliases and add MustApplyAliases

* Apply deprecated label to exposed items in /x

* Move defaults to tfbridge and move strategies to tfbridge/tokens

* Don't depend on deprecated lib

* Improve AutoAliasing doc string

* Get `make lint` working (again)

* Change the ElementStrategy interface to accept a *T (#1259)

This allows applying to an already existing mapping, such as `{Doc: ...}`.

Because it is no longer passable to call a generic strategy and then undo it (since
strategies are now handed a copy of *T), I have changed `Unmappable` to `Ignore`. If a
token is unmapped, it can be caught by our normal mechanism.

This opens up strategies to apply to arbitrary elements of `tfbride.{Resource,DataSource}Info` and to
be layered in sequence.