Fixes #1021 (#1332)
Fixes #1021

Fixes accessing RawConfig panics on seeing a nil. This bug used to breaks providers that use GetRawConfig() in diff customizers, which is a common idiom in TF. The fix is along the same lines as #1314 - we propagate the current notion of rawConfigVal to populate the structure in question. This should be much better than panic in these situations but there's more work remaining in the future to make sure the rawConfigVal is similar to what TF would pass in the same situation to avoid logical discrepancies.