Fix #1381 (#1382)
Fixes #1381 - with this change PreCheckCallback starts working the same
way for Plugin Framework resources as normal resources, that is it can
return a modify inputs map to be used in place of the original map. This
arose in the context of investigating tags secrets leaks for pulumi-aws
and in particular how tags/tagsAll behavior is different between normal
and PF based resources in the currently released version.