Separate docs source from parsing v2 (#1387)
This PR refactors the code used to acquire the markdown files we source
our docs from. The goal is to make it easy to test file acquisition and
parsing separately.

The new interface looks like this:

https://github.com/pulumi/pulumi-terraform-bridge/blob/856055a0cd3238780c9c35bbe605282f75cc2de7/pkg/tfgen/source.go#L30-L47

This PR is best reviewed as two separate commits:
- d46360529f7198cf9f3f37df3756e3308726e10e is a byte for byte copy of
the functionality that we want to refactor from `docs.go` to the newly
created `source.go`.
- 856055a0cd3238780c9c35bbe605282f75cc2de7 implements and consumes the
interface.

This PR does not have *any* effect on downstream users. It is purely a
refactor for maintainability.