Normalize first class support for nested arguments
This is purely a refactor. There are no user facing changes.

Previously, parsed TF arguments were stored in as:

```go
type entityDocs struct {
	...
	Arguments map[string]*argumentDocs
        ...
}

type argumentDocs struct {
        // The description for this argument.
	description string

	// (Optional) The names and descriptions for each argument of this argument.
	arguments map[string]string

	// Whether this argument was derived from a nested object. Used to determine
	// whether to append descriptions that have continued to the following line
        isNested bool
}
```

As a basic example, lets parse the following:

```md
* `top1` - First top level argument
* `top2` - Second top level argument
  * `nest1` - First nested argument
```

We get:

```go
Arguments: map[string]*argumentDocs{
	"top1": {
        	description: "First top level argument",
                arguments: map[string]string{},
                isNested: false,
        },
        "top2": {
        	description: "Second top level argument"
                arguments: map[string]string{
                	"nest1": "First nested argument"
		},
                isNested: false,
        },
        "nest1": {
        	description: "First nested argument",
                arguments: map[string]string{},
                isNested: true,
	},
}
```

This provided a clear parse when arguments were not nested. When arguments are nested,
there is duplication tracked by the `isNested` bit.

Nesting with `depth > 2` is possible, but never tracked correctly in the code base. It
looks like this:

```md
* `t1` - D1
  * `t2` - D2
    * `t3` - D3
```

Parsing the above (in theory) gives:

```go
Arguments: map[string]*argumentDocs{
	"t1": {
        	description: "D1",
                arguments: map[string]string{
	        	"t2": "D2",
		},
                isNested: false,
        },
        "t2": {
        	description: "D2",
                arguments: map[string]string{
	        	"t3": "D3",
		},
                isNested: true,
        },
        "t3": {
        	description: "D3",
                arguments: map[string]string{},
                isNested: true,
        },
}
```

Observe that this scheme immediately falls down when there are "any" fields with the same
name.

When we parse this example (adapted from the AWS), we get a incorrect result:

```md
* `b1` - Some field
  * `options` - Options for `b1`
    * `type` - One of "A" or "B"
* `b2` - Another field
  * `options` - Options for `b2`
    * `type` - One of "C" or "D"
```

```go
Arguments: map[string]*argumentDocs{
	"b1": {
           	description: "Some field",
        	arguments: map[string]string{
			"options": "Options for `b1`",
                },
                isNested: false,
        },
        "options": {
           	description: "Options for `b1`",
        	arguments: map[string]string{
			"type": "One of "A" or "B"",
                },
                isNested: true,
        },
        "type": {
           	description: "One of "A" or "B"",
        	arguments: map[string]string{},
                isNested: true,
        },
        "b2": {
           	description: "Some field",
        	arguments: map[string]string{
			"options": "Options for `b2`",
                },
                isNested: false,
        },
}
```

Notice that the description for `b2.options.type` is unrepresentatable in the current
schema. The scheme falls down hard. I have replaced this scheme with a single map from
path to value. This gives us a full (sparse) representation without loosing information.

This makes it much simpler to work with arguments. It also makes it much easier to iterate
on parsing, since there is an unambiguous correct parse for a given document.