Avoid false positive of nesting in bigquery_table
We had a argument:

```md
* `reference_file_schema_uri` - (Optional) When creating an external table, the user can
  provide a reference file with the table schema. This is enabled for the following formats:
  AVRO, PARQUET, ORC.
```

Since this matched a nested object regexp: "`([a-z_]+)`.*following:?", we erroneously
tagged it as a new nesting, causing us to mislabel subsequent items. The new heuristic
doesn't allow list items to anchor nestings unless the list item matches the whole
regex. This is enforced by one of our tests.