Downstream check replacement (#1442)
Implements a replacement for downstream checks.

 The intention is to allow bridge maintainers to "try out" a change on existing real-world providers. In contrast to how downstream checks used to work, this mechanism runs full build+test cycle on each provider under test. 

Specifically, a new workflow, upgrade-providers-test.yml is added that can be triggered on-demand through the Actions tab, selecting a desired bridge branch or reference. This workflow takes a dozen of real providers and upgrades them to the desired bridge reference, creating new PRs in these providers' repositories. The PRs cross-link back to the bridge PR.

The workflow needs to be used sparingly at the moment as it is expensive in consuming CI runner time.