Replace tf config names in error strings (#1531)
Should address
https://github.com/pulumi/pulumi-terraform-bridge/issues/1424

The error referred to the tf config names:
```
must provide exactly one of "api_key", "api_token" or "api_user_service_key".
```

And it's now this:
```
must provide exactly one of "cloudflare:apiKey", "cloudflare:apiToken" or "cloudflare:apiUserServiceKey".
```

I still need to do the same for pf. 
Happy to move the code around if there's a better place for it but it
looks like `check_failures.go` doesn't do what we need here. The errors
are plain strings raised from the terraform provider and don't follow
the usual validation logic in tf - looks like the upstream provider is
missing a validation line to mark the three configs as `ExactlyOneOf`.
This means that the validation doesn't happen in `CheckConfig` but
instead happens during `Configure`.

I'm happy to also raise a PR upstream if we think that'd help.