Ensure SDKv2 logs do propagate to Pulumi
Part of the problem here is context.TODO was used inside the SDKv2 provider to break the Context
object propagation chain, and as a result Logger objects setup to capture framework logs were not
passed to the code emitting the logs.

To avoid making this change breaking, ProviderWithContext interface is added side-by-side and
implemented by sdkv2 provider shim.