Additional perf instrumentation (#1534)
IN https://github.com/pulumi/pulumi-aws/issues/2987 a customer is
experiencing very long Configure and CheckConfig times but our traces
and logs provide limited visibility into what's going on.

This PR adds a little more information. This is complicated by the fact
that there's no root span for the provider and the provider process can
be killed by the engine at any moment, so finalizing the root span is
not reliable. Therefore the strategy of collecting mem stats in a
background thread and tagging them to the root span, as in pulumi/pulumi
does not work very well.

Changes:

- add spans for key sdkv2 methods GetPluginInfo Configure, CheckConfig,
Diff, Create, Update, Delete, Invoke
- all the spans sample the go memory stats for the running process and
report them as tags on the span
- since we are investigating Configure and CheckConfig specifically,
additional tags are added to show how much time is spent tin
PreConfigureCallbacks and in provider configuration validation
- the latter is also instrumented for the PF version of Configure

With these changes we have a little more visibility into PF vs SDKv2
Configure costs and have some memory stat samples to get an idea of peak
memory use for the provider. This kind of hopes that sampling memory at
the beginning of each span is sufficient; perhaps can be improved upon
by sampling every 1s for long-running spans.