Fix panics in recoverAndCoerceCtyValue (#1568)
The fix for https://github.com/pulumi/pulumi-aws/issues/3094 introduced
a new panic in https://github.com/pulumi/pulumi-aws/issues/3118 which
this PR fixes.

The issue here is that historically MakeTerraformInput returns an
interface{} but some of the later additions work more naturally with the
cty.Value domain. We used to call schema.JSONMapToStateValue prior to
3094 to recover cty.Value from interface{}, but it turns out not to be
adequate for the purpose as it does not recognize unknown sentinel.
Unfortunately the replacement code missed the implicit type coalescing
behavior of schema.JSONMapToStateValue specifically its ability to
convert silently between booleans, numbers and strings.

This PR introduces randomized tests that show that
recoverAndCoerceCtyValue is a strict improvement over
schema.JSONMapToStateValue that is on inputs where
schema.JSONMapToStateValue succeeds, recoverAndCoerceCtyValue also
succeeds and returns the same value (modulo big.Float precision). To
make the tests pass recoverAndCoerceCtyValue falls back on
schema.JSONMapToStateValue when the default path fails to produce a
scalar value of the desired type.

In the future we may consider refactoring MakeTerraformInput to target
cty.Value directly and take charge of implicit silent conversions, but
for now this change should undo the recent regression.