Add a SchemaPathToPropertyPath (#1594)
This function is currently unused in the bridge itself. It is expected
that the function will be used by providers.

See https://github.com/pulumi/pulumi-gcp/pull/1474 for an example.

---

Some tests had PP names that didn't make sense with the given schema. I
corrected them to allow round-tripability:

```patch
-			pp:       []any{"objSet", 0},
+			pp:       []any{"objSets", 0},
```

I added a test with name overrides to confirm that it still worked as
expected. Put simply, the property path []any{"listStr"} would have
never been generated by the bridge from it's schema, since
TerraformToPulumiName would have pluralized listStr to listStrs. This
worked in the context of the previous test, but lost information in the
translation.