Dont apply TF defaults before validating resource input (#1583)
This PR fixes an inconsistency in how we treat defaults versus TF.
Previously we would have applied defaults before validating input. TF
does the reverse.

Should address
https://github.com/pulumi/pulumi-terraform-bridge/issues/1546.

Tested in AWS here: https://github.com/pulumi/pulumi-aws/pull/3157


Note that this PR technically introduces a breaking change. An example
of how it'd manifest is in
[TestDefaultsAndRequiredWithValidationInteraction/CheckMissingRequiredWith](https://github.com/pulumi/pulumi-terraform-bridge/blob/5ef834fd75b7c259dacc8fb71a0cc6edfd2472bc/pkg/tfbridge/provider_test.go#L2514).

Previously we'd apply the defaults before passing to the TF provider for
validation, so this test case would have yielded no failures. This is
not what TF does - TF would fail validation on that input. We will now
match the TF behaviour and fail this input. This should be fine as the
input likely does not make sense anyway.