Refactor IgnoreChanges (#1639)
The intent here is to refactor the code without any observable changes
to make possible further extensions such as
https://github.com/pulumi/pulumi-terraform-bridge/pull/1614 .

Specifically, before this PR:

```go

type Provider interface {
        ...
	Diff(t string, s InstanceState, c ResourceConfig) (InstanceDiff, error)
}

type InstanceDiff interface {
        ...
	IgnoreChanges(ignored map[string]bool)
}

```

That is the interface is to compute a diff first, and then side-effect
to ignore changes.

After this PR we have this instead:

```go
type Provider interface {
        ...
	Diff(t string, s InstanceState, c ResourceConfig, opts DiffOptions) (InstanceDiff, error)
}
type IgnoreChanges = func() map[string]struct{}
type DiffOptions struct {
  IgnoreChanges IgnoreChanges
}
```

So the algorithm to ignore changes is specified upfront when
constructing the diff. This seems to be useful to un-constrain the
implementation of shim.Provider a little bit, specifically in #1614 that
is trying to change the internal representation of InstanceDiff
implementation in such a way that it is inconvenient to process
IgnoreChanges after the InstanceDiff is constructed already.