Improve error messages for Invalid or unknown key (#1663)
Per https://github.com/pulumi/pulumi/issues/3546 it is possible that
when a user specifies `aws:foo` config in a Pulumi program she is not
intending it to be picked up by the aws provider. This change improves
the error message emitted in this case to guide the user better.