Fixes refresh of resources with unchanged assets
Before this change, refreshing an unchanged resource which used an Asset or an Archive resulted in
polluting Pulumi state with a machine-local filename.

The simplest example this is tested on is:

    const exampleBucketObject = new aws.s3.BucketObject("exampleBucketObject", {
        key: "someobject",
        bucket: bucket.id,
        source: new pulumi.asset.FileAsset(inFile),
    });

After this change this example refreshes correctly with no changes detected.

Unfortunately this change by itself is insufficient to guarantee that these resources refresh
correctly when Pulumi-tracked state is out of date with respect to the cloud state, and further work
is required to make this correct for individual resources in each provider. For the example above,
the upstream provider for BucketObject does not implement fetching contents of the object upon Read
into a temporary file. Therefore refreshing the object in this situation updates the "etag" property
but does not guide the user to the need of reconciling the FileAsset with the updated cloud state.