Fixes mappings file naming for examples conversion (#1675)
This change is motivated by
https://github.com/pulumi/pulumi-azure/pull/1377 - it seems that the
correct way to identify a provider to the TF converter is by using the
resource prefix. This makes sure "azurerm" is used and not "azure" for
the Azure Classic prover for resource conversion, which in turn makes it
use the in-flight ProviderInfo and stop trying to download an Azure
Classic build from GitHub.