Fix example double-counting under PULUMI_CONVERT=1 (#1676)
Fixes https://github.com/pulumi/pulumi-terraform-bridge/issues/1572

Before the change examples were accidentally double-counted under
PULUMI_CONVERT=1 skewing stats. This is now fixed.

Also the interface to the coverage tracker assumed that the methods are
called in a particular order. This confused the PULUMI_CONVERT because
it shuffles the ordering of discovering and post-processing the
examples. This is now made more explicit by passing *Example pointers in
the interface to the converter.