Improve examples cache (#1717)
Fleshing out the examples cache a bit. With the introduction of
PULUMI_CONVERT=1 the build times are making large providers like AWS
very unpleasant to work with. In particular examples are being
re-rendered repeatedly across `make tfgen` and `make build_sdks` as well
as across multiple passes through the provider. This PR improves
pre-existing examples cache a bit with the following goals in mind:

- a lot more information that might be relevant to examples generation
is captured into the cache key
- so that changes in tooling versions that might affect example
rendering should invalidate the cache
- cache directories now have a metadata file explaining the details of
the cache key
- PULUMI_CONVERT=1 style conversions now support accelerating through
the cache
- this is slightly tricky as PULUMI_CONVERT=1 bulk-detects HCLs but now
needs to be aware of the cache to skip them
- cache hits should not affect example coverage tracker metrics