fix: fixes are not applied to property deprecation messages (#1765)
We have logic in `fixupPropertyReference` to replace terraform resource
references with pulumi references (i.e. `aws_s3_bucket` =>
`aws.s3.BucketV2`).

This logic is applied to the `Description` property, but it is not
applied to the separate `DeprecationMessage` property. These two
properties exist separately in the pulumi resource schema and are
combined when creating the website docs and the SDK docs.

fixes [#3028](https://github.com/pulumi/pulumi-aws/issues/3028)