Add an example schema-aware test (#1802)
Currently it is difficult to add quick unit tests that have an accurate
Pulumi Package Schema available in the provider. Adding them in the
tfbridge package runs into the limitation that these tests cannot
reference tfgen due to circular references. This change adds a new
package that works around the limitations. Unfortunately it can only
test the exported API and not the internals, but that may be sufficient
for some applications.