Root caused why ID fields were falling through to look at Attributes
docsPath was only `azure.id`. The Argument key to find, however, was `exclude.azure.id`.
This means that we kept looking for the keys "azure.id" and "id".
The latter falls through to Attribute map, which has an "id" key and thus the nested ids read "The ID of the resource".
If I hard-code "exclude" as an extra prefix, all of the buggy behavior goes away for that particular resource.
I just don't know how to get the name of the parent resource from typePath, which is DEEPLY OBNOXIOUS.