Remove the `rawNames` parameter from schema.go
`rawNames` is used to represent the distinction between traversing a Map and a Object. As
implemented, it is never correct to respect. The property should never be inherited
between functions, since the encoding of `T` does not depend on if `T` is found in
`Object<Field: T>` or `Map<T>`.

- It is already ignored in `extractSchemaInputs`.