Do not map resources/datasources specified in IgnoreMappings (#1840)
Fixes https://github.com/pulumi/pulumi-terraform-bridge/issues/1838

Previously, we applied (in bridge) automatic mapping to
resources/datasources that didn't have an associated `Tok` field, even
if they were included in `IgnoreMappings`. For providers that vendored
resources from multiple prefixes, this caused a panic.

The solution is to not provide a mapping for resources that are
specified in `IngoreMappings`. This expands the scope of
`IngoreMappings`, but does so in a way that I feel is more intuitive
then it is now. I have adjusted the comment on `IgnoreMappings` to
reflect it's new scope.