Do not fail conversion of non-HCL fenced markdown snippets
The documentation pipeline was looking for Markdown code-fenced blocks. In the case where they specify terraform or hcl
as the language, it is clear that they should be converted. However the unmarked blocks may be carrying code that needs
to be converted or else they can be carrying text that needs to be displayed with a fixed font.

This change introduces a regex-based guess that allows non-Terraform code snippets to pass through as-is.

Fixes the following AWS build warnings (107 total):

```
   1 warning: failed to convert HCL for #/types/aws:autoscaling/GroupMixedInstancesPolicyLaunchTemplateOverrideInstanceRequirements:GroupMixedInstancesPolicyLaunchTemplateOverrideInstanceRequirements/acceleratorManufacturers to csharp: <nil>: unexpected HCL snippet in Convert;
```