Reduce tfgen failing example warning verbosity (#1843)
Before this change we emitted a lot of duplicated warnings for examples
that failed to convert. This change removes the
duplication and aggregates similar errors so that when every language
conversion fails with the same error, which is
common when the failure pertains to the HCL->PCL pass of the conversion,
we get one warning only not many.