Handle provider meta better in PlanResourceChange (#1826)
fixes https://github.com/pulumi/pulumi-terraform-bridge/issues/1822 and
unlocks https://github.com/pulumi/pulumi-gcp/issues/1874

`ProviderMeta` seems to be an old experimental feature of TF which has
not picked up.

AWS does not use it and GCP uses it only for one thing - setting
UserAgent strings:
https://github.com/modular-magician/terraform-provider-google-beta/commit/6cf6c51b6b3804d6ce050885c98ae030f378bd7e

The meta we receive does not match the schema provided and does not
contain the module_name property. We should set it to null and
potentially revisit if this causes issues. I've tested it for the GCP
resource in https://github.com/pulumi/pulumi-terraform-bridge/pull/1825
and works fine.


I've opened
https://github.com/pulumi/pulumi-terraform-bridge/issues/1827 as a
follow-up to test with one of the resources which actually use the
ProviderMeta.

---------

Co-authored-by: Ian Wahbe <ian@wahbe.com>