Remove `rawNames` from `makeObjectTerraformInputs`
It never makes sense to pass `rawNames=true` to `makeObjectTerraformInputs` since that
implies we have an object type with properties, but that we should not do property name
translations. This is never the case.

This PR introduces a subtle change in behavior when walking values of unknown
type:

Before this PR, we assumed that an untyped `resource.PropertyMap` was an object (name
translation is performed) but all nested `resource.PropertyMap` values were then assumed
to be a map (name translation is not performed).

After this PR, we treat all untyped `resource.PropertyMap` values as maps. IMO this makes
more sense: the rule is we don't apply name translation unless we know a value is an
object (and can thus generate naming tables for it).

I don't expect that this change will break users since user's should never hit untyped
values. If they do, neither before or after will work consistently.

Addressing the test change:

"object_property_value" was used to test object names but never given a type. In light of
the above change, we needed to make it typed to get the expected name translation.