Breaking: Remove rawNames from `MakeTerraformOutput` & `MakeTerraformOutputs`
This has the same effect as 261df01eec2befc83b4f256578efeb09df6b3d3e, and needs the same
test case adjustment.

This is a "breaking change" to public functions, but we don't expect that these functions
will be used outside of the bridge and we have made breaking changes to them in the past (https://github.com/pulumi/pulumi-terraform-bridge/pull/1642).