Ensure validate called during tfgen (#1854)
fixes https://github.com/pulumi/pulumi-terraform-bridge/issues/1852

https://github.com/pulumi/pulumi-terraform-bridge/pull/1669 removed the
runtime calls to `InternalValidate` to speed up provider startup. This
assumed it was called during tfgen but it turns out it was not.

This introduces an explicit call to `InternalValidate` during tfgen.
Thanks to @t0yv0 for prompting the investigation and @iwahbe for helping
with the muxer.

To facilitate that, I've added `InternalValidate` to the shim Provider
interface. Technically breaking but seems unlikely to be depended on
externally.

Calling `Validate` does not do the job as that requires passing in a
valid config which is not available during tfgen.

I've added an e2e test for an sdk provider, need to add one for a pf
provider.

---------

Co-authored-by: Ian Wahbe <ian@wahbe.com>