feat: add resource urn to context in Check call
This PR adds the resource URN to the context in the `Check` method,
along with methods to set and get the URN from context.

The motivation for this change was driven by trying to implement the
suggested fix in [this
comment](https://github.com/pulumi/pulumi-aws/issues/3788#issuecomment-2051881787)
in the aws provider. We want to be able to keep some global state of
which resources have been seen by `Check`, but realized that we don't
actually get any identifying information in the `PreCheckCallback`
function.

An alternative approach would be to update the `PreCheckCallback`
signature to also contain the URN, but that would be a breaking change.

re https://github.com/pulumi/pulumi-aws/issues/3788