Exclude code choosers in SDKs
Right now, we generate code blocks for the schema that look like this:

```md
  <!--Start PulumiCodeChooser -->
  ```typescript
  // TS Code
  ```
  ```python
  # Python Code
  ```
  ...
  <!--End PulumiCodeChooser -->
```

That works great in the registry. For SDKs, we generate code that looks like this:

```python
  # <!--Start PulumiCodeChooser -->
  # ```python
  # # Python Code
  # ```
  # <!--End PulumiCodeChooser -->
```

This is not good, since our users see our PulumiCodeChooser directives directly.

This PR changes the bridge's behavior to only emit the PulumiCodeChooser directives when
we are targeting the schema.