Instrument DecodePropertyMap failures to assist root-causing refresh issues in Plugin Framework (#1920)
Investigating
https://github.com/pulumi/pulumi-terraform-bridge/issues/1919

DecodePropertyMap called from Read fails because it decoded something to
a null but a PropertyMap is expected. It appears it should never happen
because Read will not call DecodePropertyMap with a tftypes.Value=nil
but instead short-circuit the "resource not found path".

`resp.State.RemoveResource(ctx)` likewise seems to work correctly.

Testing this further, still unable to reproduce the scenario in test.
Instead, opting for adding some more debugging instrumentation around
the line that emits the error, and doing some conservative code
tightening around this code path.