fix: unknown values from PlanResourceChange are not handled correctly
When `PlanResourceChange` is enabled, the schema that is passed to
`MakeTerraformOutputs` is a `cty.Value`. We then have to convert this
into a JSON object as part of `objectFromCtyValue` before eventually
converting to Pulumi values. We were relying on a couple of functions
from other packages to do this conversion between `cty.Value` and a JSON
object which led to issues like #1932.

For example, if we have a map value like this:
```json
{
  "tagSpecifications": [
    {
      "tags": {
        "Name": "04da6b54-80e4-46f7-96ec-b56ff0331ba9"
      }
    }
    {
      "tags": {
        "Name": "Bucket Arn"
      }
    }
  ]
}
```

The first `tags` map contains an `unknown` sentinel so it will be
converted by `PlanResourceChange` into something like this:

```
{
  "tagSpecifications": [
    {
      "tags": UnknownMapValue
    }
    {
      "tags": {
        "Name": "Bucket Arn"
      }
    }
  ]
}
```

As part of the conversion we were using the `cty.Transform` function
which performs some type checking to ensure that every item in the list
is of the same type.

Since all we want from this function is to convert `cty.Value` to a JSON
object, I've taken the underlying `cty.marshal` function and modified it
to handle unknown values.

fixes #1943, re #1932