chore: catch panics when type checking logs warning (#2000)
The type checker is meant to catch user configuration errors that would
eventually cause a panic. We are currently only logging a warning so we
are not preventing the panic from happening. Even though we have the
additional warning, it is still confusing to users and some do not
correlate the warning with the panic.

This PR catches the panic when we do have type errors so that the user
will only see the warning and not the panic.

closes #1987