feat: type check config values
This PR extends the typechecker to also type check config values in
`CheckConfig`. The motivation behind this is
https://github.com/pulumi/pulumi-aws/issues/4004. In that issue the user
provides a key that is not in the allowed list and gets an opaque error
message. In Terraform the user gets a more descriptive error message, so
our behavior here is worse.

The big difference between the previous PR that is typechecking resource
inputs is in that case we were preventing panics whereas in this case
the primary motivation is to give _better_ error messages. I am still
only writing a warning for now which gives a message to the user like
this:

```console
  pulumi:providers:aws (provider):
    warning: Type checking failed:
    warning: Unexpected type at field "endpoints[0]":
               an unexpected argument "abcd" was provided
    warning: Type checking is still experimental. If you believe that a warning is incorrect,
    please let us know by creating an issue at https://github.com/pulumi/pulumi-terraform-bridge/issues.
    This will become a hard error in the future.
    error: pulumi:providers:aws resource 'chall-provider' has a problem: could not validate provider configuration: Invalid or unknown key. Examine values at 'chall-provider.endpoints'.
```

fixes https://github.com/pulumi/pulumi-aws/issues/4004